requirements.txt
asgiref==3.5.0
chardet==4.0.0
coverage==6.3.2
Django==4.0.3
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
flake8==4.0.1
lml==0.1.0
mccabe==0.6.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyexcel-io==0.6.6
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
XlsxWriter==3.0.3
